requirements.txt
```text
biopython
matplotlib